requirements.txt
# Core dependencies
langchain
langchain-community
langchain-core
langchain-google-genai>=1.0.3  # UPDATED: Required for Gemini 1.5
langchain-text-splitters
langgraph
google-generativeai>=0.5.2     # UPDATED: Required for Gemini 1.5
chromadb

# Data fetching and parsing
requests
beautifulsoup4
lxml
sec-edgar-downloader

# PDF processing
pypdf
pdfplumber

# Frontend
streamlit
plotly
pandas

# Utilities
python-dotenv
pydantic

# Testing
pytest
pytest-cov
pytest-asyncio
langchain-chroma